+++ docs/conf.py
@@ -199,7 +199,10 @@ epub_exclude_files = ['search.html']
 # -- Options for intersphinx extension ---------------------------------------
 
 # Example configuration for intersphinx: refer to the Python standard library.
-intersphinx_mapping = {'https://docs.python.org/3/': None}
+intersphinx_mapping = {
+    'https://docs.python.org/3/': None,
+    'https://2.python-requests.org/en/master/': None
+}
 
 # -- Options for todo extension ----------------------------------------------